Reworked hamiltonian.py again
This is a simpler and more easily extendable interface.

Also, I reworked the thermal_state method and switched to using it
instead of ground_state to initialize dynamics. I think this is a more
sensible choice, and it will also facilitate writing a method for
emission spectra, which should just be a very small variation of the
absorption spectra code.